// src/interfaces/IArgs.ts
export interface IArgs {
	rootDirectory?: string | boolean;
	fileExtension?: string;
	typescript?: boolean;
	redux?: boolean;
	command?: string;
	type?: any;
	resourceName?: string;
	directory?: string;
	resourceDisplayName?: string;
	resourceLink?: string;
	path?: string;
	content?: any;
	imports?: any;
	components?: any;
	isRoute?: boolean;
}